Moved NotesAdapter to other package.
